--release option breaks fix for #88
+ test isolation: clear cached jre with superimposed eea
+ move test to more likely test class
